refactor: 패키지 구조 변경, admin 도메인 분리
도메인별 패키지로 패키지 구조 분리, productController에 포함되어있던 admin 페이지를 별도 domain으로 분리
