+++ src/main/java/gift/doamin/user/dto/LoginForm.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.doamin.user.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
