+++ src/main/java/gift/doamin/user/dto/SignUpForm.java
@@ -1,5 +1,6 @@
-package gift;
+package gift.doamin.user.dto;
 
+import gift.doamin.user.entity.UserRole;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
