+++ src/main/java/gift/doamin/user/repository/UserRepository.java
@@ -1,4 +1,6 @@
-package gift;
+package gift.doamin.user.repository;
+
+import gift.doamin.user.entity.User;
 
 public interface UserRepository {
     void save(User user);
