+++ src/main/java/gift/doamin/user/service/AuthService.java
@@ -1,5 +1,11 @@
-package gift;
+package gift.doamin.user.service;
 
+import gift.doamin.user.entity.User;
+import gift.doamin.user.entity.UserRole;
+import gift.doamin.user.dto.LoginForm;
+import gift.doamin.user.dto.SignUpForm;
+import gift.doamin.user.repository.UserRepository;
+import gift.global.JwtProvider;
 import java.util.HashMap;
 import java.util.Map;
 import org.springframework.http.HttpStatus;
